app/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pydantic==2.7.4
pydantic-settings==2.3.4
python-multipart==0.0.9
numpy==1.26.4
Pillow==10.4.0
scipy==1.11.4
